requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
langgraph==0.2.34
pydantic==2.8.2
python-dotenv==1.2.2
qrcode[pil]==8.2
telethon==1.42.0
langchain-gigachat==0.3.12
requests==2.32.3
